Parallelize tests and cleanup some cruft
Use pytest-xdist to run tests in parallel.  This is especially relevant
for the remote tests.

Also now run the linter on the tests, since a simple lint error could be
a big time sink given how long these take to run (order of magnitude 10
minutes).

AWS returns timeouts and connection resets if I try to do too many
things in parallel, but that's a future investigation.  For now I only
set the number of cores to 2 for that test.

There were also some issues with the way firewalls were destroyed and
the way instances were discovered in GCE that caused some test failures,
so this commit fixes those.