Refactor to stabilize and simplify API
This is a big refactor with a lot of changes, with the goal of
stabilizing the API to something more coherent and deliberate.

To this end, this makes a bunch of changes (which are updated in the
README).  The best way to review this is to read that off of this
branch, since so much has changed and moved around that the diff doesn't
say much.

The big changes:

- Rename "client.instances" to "client.service" everywhere.
- Remove "client.subnetwork" entirely from the API.
- Make all functions return objects found in "butter/types/common.py".
- Add a "mock-aws" provider for test usage.
- Allow actually passing instance "count" to both GCE and AWS.
- Make the functions take objects instead of strings in some cases.
- Add more getting started documentation to the README.

Fixes: https://github.com/sverch/butter/issues/35