Fix logging configuration to be heirarchical
Previously, using the "cloudless.log" logger meant that any
configuration there didn't actually override the loggers in the sub
packages.  This change makes it so that the logger is actually the name
of the module I'm logging for, not suffixed by ".log" just because the
setup lives in log.py.