WIP Trying to get a model for plan
This is modeling the state of a project, which will help with a few
things, mostly plan but also providing a layer of abstraction and
validation (possibly).

This has an AWS specific model, but that seems like the WRONG idea.  The
model should be flexible enough to be able to map to that, like an
expressive intermediate representation.