Put temporary state in a .cloudless directory
This is for the image build and the state testing commands.  Before this
change they would put temporary keys and the JSON file in the top level
directory, which was non ideal in a lot of ways.

Ideally, this would have smarter namespacing (perhaps on the profile)
but this is a step in that direction.