Set AWS profile in cloudless profile
This allows a Cloudless profile to use a specific AWS profile.

A lot needed to be changed because the credentials object was threaded
through everywhere and didn't need to be as I decided to implement this.

Downside is that this sets the profile for boto globally if you use
cloudless and then boto, but this is going to require some major
refactoring anyway.